requirements.txt
absl-py==0.11.0
albumentations==0.5.1
cachetools==4.1.1
certifi==2020.11.8
chainer==7.7.0
chardet==3.0.4
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.21
dataclasses==0.8
decorator==4.4.2
filelock==3.0.12
fsspec==0.8.4
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
graphviz==0.8.4
grpcio==1.33.2
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.1.0
joblib==0.17.0
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
ml-collections==0.1.0
mxnet==1.7.0.post1
networkx==2.5
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
opencv-python-headless==4.4.0.46
pandas==1.1.4
Pillow==8.0.1
pkg-resources==0.0.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@4d5343c398d75d7ebae34f51a47769084ba3a613
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.0.8
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
Shapely==1.7.1
six==1.15.0
sklearn==0.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
threadpoolctl==2.1.0
tifffile==2020.9.3
timm==0.3.2
torch==1.7.0
torchutils==0.0.4
torchvision==0.8.1
tqdm==4.53.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0
